dom/workers/test/test_WorkerDebugger.isFrozen.xul

<?xml version="1.0"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="Test for WorkerDebugger.isFrozen"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript" src="dom_worker_helper.js"/>

  <script type="application/javascript">
  <![CDATA[

    const CACHE_SUBFRAMES = "browser.sessionhistory.cache_subframes";
    const MAX_TOTAL_VIEWERS = "browser.sessionhistory.max_total_viewers";

    const IFRAME1_URL = "WorkerDebugger.isFrozen_iframe1.html";
    const IFRAME2_URL = "WorkerDebugger.isFrozen_iframe2.html";

    const WORKER1_URL = "WorkerDebugger.isFrozen_worker1.js";
    const WORKER2_URL = "WorkerDebugger.isFrozen_worker2.js";

    function test() {
      Task.spawn(function* () {
        SimpleTest.waitForExplicitFinish();

        var oldMaxTotalViewers = SpecialPowers.getIntPref(MAX_TOTAL_VIEWERS);

        SpecialPowers.setBoolPref(CACHE_SUBFRAMES, true);
        SpecialPowers.setIntPref(MAX_TOTAL_VIEWERS, 10);

        let iframe = $("iframe");

        let promise = waitForMultiple([
          waitForRegister(WORKER1_URL),
          waitForWindowMessage(window, "ready"),
        ]);
        iframe.src = IFRAME1_URL;
        let [dbg1] = yield promise;
        is(dbg1.isClosed, false,
           "debugger for worker on page 1 should not be closed");
        is(dbg1.isFrozen, false,
           "debugger for worker on page 1 should not be frozen");

        promise = waitForMultiple([
          waitForDebuggerFreeze(dbg1),
          waitForRegister(WORKER2_URL),
          waitForWindowMessage(window, "ready"),
        ]);
        iframe.src = IFRAME2_URL;
        let [_, dbg2] = yield promise;
        is(dbg1.isClosed, false,
           "debugger for worker on page 1 should not be closed");
        is(dbg1.isFrozen, true,
           "debugger for worker on page 1 should be frozen");
        is(dbg2.isClosed, false,
           "debugger for worker on page 2 should not be closed");
        is(dbg2.isFrozen, false,
           "debugger for worker on page 2 should not be frozen");

        promise = Promise.all([
          waitForDebuggerFreeze(dbg2),
          waitForDebuggerThaw(dbg1),
        ]);
        iframe.contentWindow.history.back();
        yield promise;
        is(dbg1.isClosed, false,
           "debugger for worker on page 1 should not be closed")
        is(dbg1.isFrozen, false,
           "debugger for worker on page 1 should not be frozen");
        is(dbg2.isClosed, false,
           "debugger for worker on page 2 should not be closed");
        is(dbg2.isFrozen, true,
           "debugger for worker on page 2 should be frozen");

        SpecialPowers.clearUserPref(CACHE_SUBFRAMES);
        SpecialPowers.setIntPref(MAX_TOTAL_VIEWERS, oldMaxTotalViewers);

        SimpleTest.finish();
      });
    }

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
    <iframe id="iframe"></iframe>
  </body>
  <label id="test-result"/>
</window>
